Consuming / and ? as regular characters in find
The previous handling of f/ T? did not cover the cases where those
motions were part of an action. Fixing cf, dT, yF, and various g*t
commands when followed by a / or ? character.

The unittests included assume the vim-incompatible not-moving cursor
when yanking backwards. They can be changed easily if kwrite-devel
settles on something different later on.